add export for mapper function
