+++ packages/apollos-ui-connected/src/HorizontalContentCardConnected/index.js
@@ -1,29 +1,5 @@
-import React, { memo } from 'react';
-import PropTypes from 'prop-types';
+import HorizontalContentCardConnected from './HorizontalContentCardConnected';
 
-import ContentCardConnected from '../ContentCardConnected';
-
-import horizontalContentCardComponentMapper from './horizontalContentCardComponentMapper';
-
-const HorizontalContentCardConnected = memo(
-  ({ Component, isLoading, ...otherProps }) => (
-    <ContentCardConnected
-      Component={Component}
-      isLoading={isLoading}
-      {...otherProps}
-    />
-  )
-);
-
-HorizontalContentCardConnected.propTypes = {
-  Component: PropTypes.func,
-  isLoading: PropTypes.bool,
-};
-
-HorizontalContentCardConnected.defaultProps = {
-  Component: horizontalContentCardComponentMapper,
-};
-
-HorizontalContentCardConnected.displayName = 'HorizontalContentCardConnected';
+export horizontalContentCardComponentMapper from './horizontalContentCardComponentMapper';
 
 export default HorizontalContentCardConnected;
